remove old gallery code and rename package
