--- app/src/main/java/dev/leonlatsch/photok/cgallery/ui/GalleryFragment.kt
@@ -1,79 +1,0 @@
-/*
- *   Copyright 2020-2024 Leon Latsch
- *
- *   Licensed under the Apache License, Version 2.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- */
-
-package dev.leonlatsch.photok.cgallery.ui
-
-import android.content.res.Configuration
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.ui.platform.ComposeView
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.viewModels
-import androidx.navigation.fragment.findNavController
-import dagger.hilt.android.AndroidEntryPoint
-import dev.leonlatsch.photok.cgallery.ui.compose.GalleryScreen
-import dev.leonlatsch.photok.cgallery.ui.navigation.GalleryNavigator
-import dev.leonlatsch.photok.imageloading.compose.LocalEncryptedImageLoader
-import dev.leonlatsch.photok.other.extensions.launchLifecycleAwareJob
-import dev.leonlatsch.photok.settings.data.Config
-import javax.inject.Inject
-
-@AndroidEntryPoint
-class GalleryFragment : Fragment() {
-
-    private val viewModel: GalleryViewModel by viewModels()
-
-    @Inject
-    lateinit var navigator: GalleryNavigator
-
-    @Inject
-    lateinit var config: Config
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View = ComposeView(requireContext()).apply {
-        setContent {
-            CompositionLocalProvider(
-                LocalEncryptedImageLoader provides viewModel.encryptedImageLoader
-            ) {
-                GalleryScreen(viewModel)
-            }
-        }
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        launchLifecycleAwareJob {
-            viewModel.eventsFlow.collect { event ->
-                navigator.navigate(event, findNavController(), this)
-            }
-        }
-
-        viewModel.checkForNewFeatures()
-    }
-
-    override fun onConfigurationChanged(newConfig: Configuration) {
-        super.onConfigurationChanged(newConfig)
-
-        viewModel.onConfigurationChanged()
-    }
-}
