+++ app/src/main/java/dev/leonlatsch/photok/gallery/ui/GalleryUiEvent.kt
@@ -14,15 +14,14 @@
  *   limitations under the License.
  */
 
-package dev.leonlatsch.photok.cgallery.ui.navigation
+package dev.leonlatsch.photok.gallery.ui
 
-import dev.leonlatsch.photok.model.database.entity.Photo
-
-sealed interface GalleryNavigationEvent {
-    data class OpenPhoto(val photoUUID: String) : GalleryNavigationEvent
-    data object OpenImportMenu : GalleryNavigationEvent
-
-    data class StartDeleteDialog(val photosToDelete: List<Photo>) : GalleryNavigationEvent
-    data class StartExportDialog(val photosToExport: List<Photo>) : GalleryNavigationEvent
-    data object ShowNewFeaturesDialog : GalleryNavigationEvent
+sealed interface GalleryUiEvent {
+    data class PhotoClicked(val item: PhotoTile) : GalleryUiEvent
+    data class PhotoLongPressed(val item: PhotoTile) : GalleryUiEvent
+    object OpenImportMenu : GalleryUiEvent
+    object CancelMultiSelect : GalleryUiEvent
+    object SelectAll : GalleryUiEvent
+    object OnDelete : GalleryUiEvent
+    object OnExport : GalleryUiEvent
 }
