+++ app/src/main/java/dev/leonlatsch/photok/gallery/ui/GalleryUiStateFactory.kt
@@ -14,7 +14,7 @@
  *   limitations under the License.
  */
 
-package dev.leonlatsch.photok.cgallery.ui
+package dev.leonlatsch.photok.gallery.ui
 
 import dev.leonlatsch.photok.model.database.entity.Photo
 import javax.inject.Inject
