+++ app/src/main/java/dev/leonlatsch/photok/gallery/ui/menu/DeleteBottomSheetDialogFragment.kt
@@ -1,5 +1,5 @@
 /*
- *   Copyright 2020-2021 Leon Latsch
+ *   Copyright 2020-2024 Leon Latsch
  *
  *   Licensed under the Apache License, Version 2.0 (the "License");
  *   you may not use this file except in compliance with the License.
