+++ app/src/main/res/navigation/main_nav_graph.xml
@@ -163,7 +163,7 @@
     </fragment>
     <fragment
         android:id="@+id/cgalleryFragment"
-        android:name="dev.leonlatsch.photok.cgallery.ui.GalleryFragment"
+        android:name="dev.leonlatsch.photok.gallery.ui.GalleryFragment"
         android:label="GalleryFragment">
         <action
             android:id="@+id/action_cgalleryFragment_to_imageViewerFragment"
